# sample/mongoose-app/openapi/api-docs.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/account/registration": {
            "post": {
                "operationId": "AccountController_registration",
                "parameters": [
                    {
                        "name": "X-RECAPTCHA-TOKEN",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "recaptcha_token",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterUserDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Account"
                ]
            }
        },
        "/account/identifier-available": {
            "get": {
                "operationId": "AccountController_identifierAvailable",
                "parameters": [
                    {
                        "name": "field",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "value",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-RECAPTCHA-TOKEN",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "recaptcha_token",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatusResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Account"
                ]
            }
        },
        "/account/send-verify-email": {
            "post": {
                "operationId": "AccountController_sendVerifyEmail",
                "parameters": [
                    {
                        "name": "X-RECAPTCHA-TOKEN",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "recaptcha_token",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatusResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShortErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShortErrorResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Account"
                ],
                "security": [
                    {
                        "basic": []
                    },
                    {
                        "bearer": []
                    },
                    {
                        "oauth2": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/account/verify-email": {
            "post": {
                "operationId": "AccountController_verifyEmail",
                "parameters": [
                    {
                        "name": "X-RECAPTCHA-TOKEN",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "recaptcha_token",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifyEmailDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatusResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Account"
                ]
            },
            "get": {
                "operationId": "AccountController_verifyEmailGet",
                "parameters": [
                    {
                        "name": "token",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-RECAPTCHA-TOKEN",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "recaptcha_token",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatusResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Account"
                ]
            }
        },
        "/account/send-reset-password": {
            "post": {
                "operationId": "AccountController_sendResetPassword",
                "parameters": [
                    {
                        "name": "X-RECAPTCHA-TOKEN",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "recaptcha_token",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendResetPasswordDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatusResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Account"
                ]
            }
        },
        "/account/reset-password": {
            "post": {
                "operationId": "AccountController_resetPassword",
                "parameters": [
                    {
                        "name": "X-RECAPTCHA-TOKEN",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "recaptcha_token",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetPasswordDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatusResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Account"
                ]
            },
            "get": {
                "operationId": "AccountController_resetPasswordGet",
                "parameters": [
                    {
                        "name": "token",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "minLength": 6,
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-RECAPTCHA-TOKEN",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "recaptcha_token",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatusResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Account"
                ]
            }
        },
...
    },
    "info": {
        "title": "Ivy Backend DEV",
        "description": "<small><a href=\"/api-docs-json\" target=\"_blank\">api-docs.json</a></small><br><p>Ivy backend API service</p>",
        "version": "0.1.0",
        "contact": {}
    },
...
    "servers": [
        {
            "url": "http://127.0.0.1:4000",
            "description": "Default server"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
            "ShortErrorResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "number"
                    }
                },
                "required": [
                    "message",
                    "code"
                ]
            },
...
            "ValidationError": {
                "type": "object",
                "properties": {
                    "value": {
                        "oneOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "number"
                            },
                            {
                                "type": "boolean"
                            }
                        ]
                    },
                    "constraints": {
                        "type": "object",
                        "additionalProperties": {
                            "allOf": [
                                {
                                    "type": "string"
                                }
                            ]
                        }
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    },
                    "property": {
                        "type": "string"
                    }
                },
                "required": [
                    "value",
                    "constraints",
                    "property"
                ]
            },
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "reason": {
                        "oneOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        ]
                    },
                    "timestamp": {
                        "type": "string"
                    },
                    "path": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "number"
                    }
                },
                "required": [
                    "reason",
                    "timestamp",
                    "path",
                    "message",
                    "code"
                ]
            },
            "StatusResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "required": [
                    "success"
                ]
            },
            "FileMeta": {
                "type": "object",
                "properties": {
                    "mimeType": {
                        "type": "string"
                    },
                    "size": {
                        "type": "number"
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    }
                }
            },
            "File": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "string"
                    },
                    "originalName": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/FileMeta"
                    }
                },
                "required": [
                    "data"
                ]
            },
...
            "Application": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "scheduledAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "project": {
                        "$ref": "#/components/schemas/Project"
                    },
                    "reviewers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "title",
                    "project",
                    "category"
                ]
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "User",
                                "Admin"
                            ]
                        }
                    },
                    "consent": {
                        "type": "boolean"
                    },
                    "verified": {
                        "type": "boolean"
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "loginAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "logoutAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "avatar": {
                        "$ref": "#/components/schemas/File"
                    },
                    "projects": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Project"
                        }
                    },
                    "reviewedApps": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Application"
                        }
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "id"
                ]
            },
...
            "Project": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "owner": {
                        "$ref": "#/components/schemas/User"
                    },
                    "plan": {
                        "$ref": "#/components/schemas/Plan"
                    },
                    "applications": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Application"
                        }
                    },
                    "documents": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/File"
                        }
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "plan"
                ]
            },
...
            "FileDto": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "string"
                    },
                    "originalName": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 4096
                    }
                },
                "required": [
                    "data"
                ]
            },
...
            "RegisterUserDto": {
                "type": "object",
                "properties": {
                    "consent": {
                        "type": "boolean"
                    },
                    "firstName": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "lastName": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "email": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "password": {
                        "type": "string",
                        "minLength": 6
                    },
                    "avatar": {
                        "$ref": "#/components/schemas/FileDto"
                    }
                },
                "required": [
                    "consent",
                    "firstName",
                    "lastName",
                    "email",
                    "password"
                ]
            },
            "VerifyEmailDto": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    }
                },
                "required": [
                    "token"
                ]
            },
            "SendResetPasswordDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    }
                },
                "required": [
                    "email"
                ]
            },
            "ResetPasswordDto": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string",
                        "minLength": 6
                    }
                },
                "required": [
                    "token",
                    "password"
                ]
            },
...
        }
    }
}